requirements.txt
anyio==3.7.1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.5.7
chardet==3.0.4
charset-normalizer==3.2.0
exceptiongroup==1.1.2
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
pip==21.1.2
python-telegram-bot==13.15
pytz==2023.3
requests==2.31.0
rfc3986==1.5.0
setuptools==57.0.0
six==1.16.0
sniffio==1.3.0
tornado==6.1
tzlocal==5.0.1
urllib3==2.0.3
wheel==0.36.2
